feat(ui): add official MCP logo as custom icon
- Add McpIcon component with official MCP logo SVG
- Rename capability-icons.ts to .tsx for JSX support
- Update MCP capability to use "mcp" icon name

Logo source: https://github.com/modelcontextprotocol/modelcontextprotocol
